Restructure Flask application to use Blueprints for improved modularity and maintainability.
- Moved existing routes to `plc.py` under `audit` Blueprint.
- Updated `base.html` and other templates to reference Blueprint's routes.
- Adjusted `form.html` form submission route to align with the Blueprint structure.
- Added `create_app` factory function for Flask app creation.
