+++ presentations/templates/form.html
@@ -22,7 +22,7 @@
 
 <div class="row">
   <div class="col-md-6">
-    <form method="post" action="{{ url_for('form') }}">
+    <form method="post" action="{{ url_for('audit.form') }}">
       <div class="mb-3">
         <label for="name" class="form-label">Name *</label>
         <input type="text" class="form-control" id="name" name="name" required value="{{ form_data.name if form_data and form_data.name }}">
